fix(task_filters): async conditions and error handling on task filters
* Updated task filtering to support asynchronous conditions, improving the accuracy of task retrieval.
* Adjusted the project tasks board and sectioned list views to utilize the new filtering logic.
* Improved the user queries to fetch assigned user IDs for tasks, enhancing the async filtering performance.
* Simplified tasksAndParentsByProjectFilteredProvider logic ensuring consistency

+++ lib/services/data/projects/repositories/projects_repository.dart
@@ -99,7 +99,7 @@ class SearchProjectResult {
     return SearchProjectResult(
       id: json['id'],
       name: json['name'],
-      similarityScore: json['similarity_score'],
+      similarityScore: json['similarity_score'].toDouble(),
     );
   }
 }
